% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMCMC.R
\name{addMCMC}
\alias{addMCMC}
\title{func_name()}
\usage{
addMCMC(model_script, analysis_name, output_folder, generations = 10000,
  burnin = 250, tuning = 100)
}
\arguments{
\item{param_1}{Insert description}
}
\description{
Insert description
}
\keyword{mrRevBayes}
