% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMCMCMC.R
\name{addMCMCMC}
\alias{addMCMCMC}
\title{func_name()}
\usage{
addMCMCMC(model_script, analysis_name, output_folder, nchains = 4,
  deltaheat = 0.2, generations = 10000, burnin = 250, tuning = 100)
}
\arguments{
\item{param_1}{Insert description}
}
\description{
Insert description
}
\keyword{mrRevBayes}
