% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{configSubsetCheck}
\alias{configSubsetCheck}
\title{func_name()}
\usage{
configSubsetCheck(config_mat)
}
\arguments{
\item{param_1}{Insert description}
}
\description{
Insert description
}
\keyword{mrRevBayes}
