% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{getBayesBlock}
\alias{getBayesBlock}
\title{getBayesBlock()}
\usage{
getBayesBlock(nexus_file)
}
\arguments{
\item{nexus_file}{Relative path to a nexus file with a Bayes block in it}
}
\description{
Returns the Bayes block from a nexus file
}
\keyword{mrRevBayes}
