% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{getSubsets}
\alias{getSubsets}
\title{getSubsets()}
\usage{
getSubsets(bayes_block)
}
\arguments{
\item{bayes_block}{str - Takes the output from getBayesBlock()}
}
\description{
Returns the subsets specified by a Bayes block
}
\keyword{mrRevBayes}

