% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioned_model.R
\name{gtr_model_2}
\alias{gtr_model_2}
\title{gtr_model_2()}
\usage{
gtr_model_2(ix)
}
\arguments{
\item{ix}{int/str - Current partition's index number}
}
\description{
Used in partitioned_model() for GTR models
}
\keyword{mrRevBayes}
