% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uniform_model.R
\name{hky_model}
\alias{hky_model}
\title{hky_model()}
\usage{
hky_model(gamma_pinv, subset_name = "")
}
\arguments{
\item{gamma_pinv}{Boolean vector from uniform_model()}

\item{subset_name}{Deprecated; this paramater is needed for an old partitioned_model().}
}
\description{
Writes code for the HKY model for uniform_model()
}
\keyword{mrRevBayes}

