% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partitioned_model.R
\name{hky_model_2}
\alias{hky_model_2}
\title{hky_model_2()}
\usage{
hky_model_2(ix)
}
\arguments{
\item{ix}{int/str - Current partition's index number}
}
\description{
Used in partitioned_model() for HKY models
}
\keyword{mrRevBayes}

