% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partitioned_model.R
\name{jc_model_2}
\alias{jc_model_2}
\title{jc_model_2()}
\usage{
jc_model_2(ix)
}
\arguments{
\item{ix}{int/str - Current partition's index number}
}
\description{
Currently has no purpose
}
\keyword{mrRevBayes}

