% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partitioned_model.R
\name{k80_model_2}
\alias{k80_model_2}
\title{k80_model_2()}
\usage{
k80_model_2(ix)
}
\arguments{
\item{ix}{int/str - Current partition's index number}
}
\description{
Used in partitioned_model() for K80 models
}
\keyword{mrRevBayes}

