% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_model.R
\name{likelihood_function}
\alias{likelihood_function}
\title{func_name()}
\usage{
likelihood_function(gamma_pinv, subset_name = FALSE)
}
\arguments{
\item{param_1}{Insert description}
}
\description{
Insert description
}
\keyword{mrRevBayes}
