% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{link_params}
\alias{link_params}
\title{link_params()}
\usage{
link_params(link_string)
}
\arguments{
\item{link_string}{str - Takes in a linking line from a Bayes block}
}
\description{
Classifies whether or not their exists linking lines
}
\keyword{mrRevBayes}

