% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{linkMatrix}
\alias{linkMatrix}
\title{func_name()}
\usage{
linkMatrix(link_string, bayes_links, headers = TRUE, main = TRUE,
  blank = FALSE)
}
\arguments{
\item{param_1}{Insert description}
}
\description{
Insert description
}
\keyword{mrRevBayes}
