% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{mrRevBayes}
\alias{mrRevBayes}
\title{mrRevBayes()}
\usage{
mrRevBayes(nexus_file)
}
\arguments{
\item{nexus_file}{Relative path to nexus file with a Bayes block in it}
}
\description{
Takes a nexus file and makes a scheme
}
\keyword{mrRevBayes}
