% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{nst_name}
\alias{nst_name}
\title{nst_name()}
\usage{
nst_name(nst_num)
}
\arguments{
\item{nst_num}{int/str - Values include NA, 1, 2, or 6}
}
\description{
Translates nst to a model name
}
\keyword{mrRevBayes}

