% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioned_model.R
\name{partitioned_model}
\alias{partitioned_model}
\title{func_name()}
\usage{
partitioned_model(scheme_csv, subsets_folder, name)
}
\arguments{
\item{param_1}{Insert description}
}
\description{
Insert description
}
\keyword{mrRevBayes}
