% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{rate_logical}
\alias{rate_logical}
\title{rate_logical()}
\usage{
rate_logical(rates)
}
\arguments{
\item{rates}{str - Values include "equal", "gamma", "propinv", "invgamma"}
}
\description{
Translates the ASRV information
}
\keyword{mrRevBayes}

