% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{readLinks}
\alias{readLinks}
\title{readLinks()}
\usage{
readLinks(bayes_block)
}
\arguments{
\item{bayes_block}{Output from getBayesBlock()}
}
\description{
Reads in links within mrRevBayes()
}
\keyword{mrRevBayes}
