% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{readLsets}
\alias{readLsets}
\title{readLsets()}
\usage{
readLsets(bayes_block, num_subsets)
}
\arguments{
\item{bayes_block}{Output from getBayesBlock()}

\item{num_subsets}{int/str - How many subsets specified in the Bayes block}
}
\description{
Reads in lsets within mrRevBayes()
}
\keyword{mrRevBayes}

