% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{readPriors}
\alias{readPriors}
\title{readPriors()}
\usage{
readPriors(bayes_block)
}
\arguments{
\item{bayes_block}{Output from getBayesBlock()}
}
\description{
Reads in priors within mrRevBayes()
}
\keyword{mrRevBayes}
