% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{simplifyRows}
\alias{simplifyRows}
\title{simplifyRows()}
\usage{
simplifyRows(this_row)
}
\arguments{
\item{this_row}{Row from the configuration file scheme}
}
\description{
Reads in priors within mrRevBayes()
}
\keyword{mrRevBayes}

