% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partitioned_model.R
\name{sym_model_2}
\alias{sym_model_2}
\title{sym_model_2()}
\usage{
sym_model_2(ix)
}
\arguments{
\item{ix}{int/str - Current partition's index number}
}
\description{
Used in partitioned_model() for SYM models
}
\keyword{mrRevBayes}

