% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_model.R
\name{uniform_model}
\alias{uniform_model}
\title{func_name()}
\usage{
uniform_model(scheme_csv, nexus_file, name)
}
\arguments{
\item{param_1}{Insert description}
}
\description{
Insert description
}
\keyword{mrRevBayes}
