% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mrRevBayes.R
\name{configSubsetCheck}
\alias{configSubsetCheck}
\title{configSubsetCheck()}
\usage{
configSubsetCheck(config_mat)
}
\arguments{
\item{config_mat}{matrix - NOT A RELATIVE PATH}
}
\description{
Cleans up configuration files by separating columns that are oddly together and deleting duplicate columns
}
\keyword{mrRevBayes}

