% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_model.R
\name{gtr_model}
\alias{gtr_model}
\title{gtr_model()}
\usage{
gtr_model(gamma_pinv, subset_name = "")
}
\arguments{
\item{gamma_pinv}{Boolean vector from uniform_model()}

\item{subset_name}{Deprecated; this paramater is needed for an old partitioned_model().}
}
\description{
Writes code for the GTR model for uniform_model()
}
\keyword{mrRevBayes}
