% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_model.R
\name{hky_model}
\alias{hky_model}
\title{func_name()}
\usage{
hky_model(gamma_pinv, subset_name = "")
}
\arguments{
\item{param_1}{Insert description}
}
\description{
Insert description
}
\keyword{mrRevBayes}
