% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uniform_model.R
\name{likelihood_function}
\alias{likelihood_function}
\title{likelihood_function()}
\usage{
likelihood_function(gamma_pinv, subset_name = FALSE)
}
\arguments{
\item{gamma_pinv}{Boolean vector from uniform_model()}

\item{subset_name}{Deprecated; this paramater is needed for an old partitioned_model().}
}
\description{
Used with uniform_model(). This function writes the dnPhyloCTMC lines for Rev scripts.
}
\keyword{mrRevBayes}

