% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partitioned_model.R
\name{partitioned_model}
\alias{partitioned_model}
\title{partitioned_model()}
\usage{
partitioned_model(scheme_csv, subsets_folder)
}
\arguments{
\item{scheme_csv}{.csv - Relative path to the configuration file}

\item{subsets_folder}{path - Relative path to folder that contains nexus subset files}
}
\description{
Outputs Rev script given a RevBayes configuration file
}
\keyword{mrRevBayes}

