% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_model.R
\name{uniform_model}
\alias{uniform_model}
\title{uniform_model()}
\usage{
uniform_model(scheme_csv, nexus_file)
}
\arguments{
\item{scheme_csv}{.csv - Relative path to the configuration file}

\item{nexus_file}{.nex - Relative path to the nexus file with the Bayes block in it}
}
\description{
Outputs Rev script given a RevBayes configuration file
}
\keyword{mrRevBayes}
